Fix bug with sprites not appearing on left and top side of window

+++ src/gameboy/PPU/Sprite.h
@@ -29,8 +29,8 @@ public:
 
     bool coversLine(uint8_t line, unsigned int objectSize) const;
 
-    uint8_t getX() const;
-    uint8_t getY() const;
+    int getX() const;
+    int getY() const;
     uint8_t getTileID(uint8_t lcdY) const;
     uint8_t getTileX(uint8_t lcdX) const;
     uint8_t getTileY(uint8_t lcdY) const;
